refactor(front): move prismaQuery types to services folder
